Update
Banlist Update
Dashboard, include cold stakes
fix typo in settingsbittoolwidget
fix stakingbalance calculation
clear addresslable when address is changed
