+++ src/consensus/tx_verify.cpp
@@ -5,6 +5,7 @@
 #include "tx_verify.h"
 
 #include "consensus/consensus.h"
+#include "consensus/zerocoin_verify.h"
 #include "main.h"
 #include "script/interpreter.h"
