+++ src/consensus/zerocoin_verify.h
@@ -1,0 +1,28 @@
+// Copyright (c) 2020 The PIVX developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef PIVX_CONSENSUS_ZEROCOIN_VERIFY_H
+#define PIVX_CONSENSUS_ZEROCOIN_VERIFY_H
+
+#include "consensus/consensus.h"
+#include "main.h"
+#include "script/interpreter.h"
+#include "zepgchain.h"
+
+/** Context-independent validity checks */
+bool CheckZerocoinSpend(const CTransaction& tx, bool fVerifySignature, CValidationState& state, bool fFakeSerialAttack = false);
+// Fake Serial attack Range
+bool isBlockBetweenFakeSerialAttackRange(int nHeight);
+// Public coin spend
+bool CheckPublicCoinSpendEnforced(int blockHeight, bool isPublicSpend);
+int CurrentPublicCoinSpendVersion();
+bool CheckPublicCoinSpendVersion(int version);
+bool ContextualCheckZerocoinSpend(const CTransaction& tx, const libzerocoin::CoinSpend* spend, int nHeight, const uint256& hashBlock);
+bool ContextualCheckZerocoinSpendNoSerialCheck(const CTransaction& tx, const libzerocoin::CoinSpend* spend, int nHeight, const uint256& hashBlock);
+void AddWrappedSerialsInflation();
+bool RecalculatePIVSupply(int nHeightStart);
+bool UpdateZEPGSupply(const CBlock& block, CBlockIndex* pindex, bool fJustCheck);
+CAmount GetInvalidUTXOValue();
+
+#endif //PIVX_CONSENSUS_ZEROCOIN_VERIFY_H
