+++ src/miner.cpp
@@ -569,7 +569,7 @@ void BitcoinMiner(CWallet* pwallet, bool fProofOfStake)
 {
     LogPrintf("EPGCMiner started\n");
     SetThreadPriority(THREAD_PRIORITY_LOWEST);
-    RenameThread("epgc-miner");
+    util::ThreadRename("epgc-miner");
     const int64_t nSpacingMillis = Params().GetConsensus().nTargetSpacing * 1000;
 	const int last_pow_block = Params().GetConsensus().height_last_PoW;
