+++ src/obfuscation.cpp
@@ -587,7 +587,7 @@ void ThreadCheckObfuScationPool()
     if (fLiteMode) return; //disable all Obfuscation/Masternode related functionality
 
     // Make this thread recognisable as the wallet flushing thread
-    RenameThread("epgc-obfuscation");
+    util::ThreadRename("epgc-obfuscation");
     LogPrintf("Masternodes thread started\n");
 
     unsigned int c = 0;
