+++ src/qt/epgc/settings/settingssignmessagewidgets.cpp
@@ -61,9 +61,7 @@ SettingsSignMessageWidgets::SettingsSignMessageWidgets(EPGCGUI* _window, QWidget
     ui->labelSubtitleMessage->setText(tr("Message"));
     ui->labelSubtitleMessage->setProperty("cssClass", "text-title");
 
-#if QT_VERSION >= 0x050300
     ui->messageIn_SM->setPlaceholderText(tr("Write message"));
-#endif
     ui->messageIn_SM->setProperty("cssClass","edit-primary");
     setShadow(ui->messageIn_SM);
     ui->messageIn_SM->setAttribute(Qt::WA_MacShowFocusRect, 0);
