+++ src/qt/macdockiconhandler.mm
@@ -50,9 +50,7 @@ MacDockIconHandler::MacDockIconHandler() : QObject()
     this->m_dummyWidget = new QWidget();
     this->m_dockMenu = new QMenu(this->m_dummyWidget);
     this->setMainWindow(nullptr);
-#if QT_VERSION >= 0x050200
     this->m_dockMenu->setAsDockMenu();
-#endif
     [pool release];
 }
