+++ src/wallet/walletdb.cpp
@@ -847,7 +847,7 @@ DBErrors CWalletDB::ZapWalletTx(CWallet* pwallet, std::vector<CWalletTx>& vWtx)
 void ThreadFlushWalletDB(const std::string& strFile)
 {
     // Make this thread recognisable as the wallet flushing thread
-    RenameThread("epgc-wallet");
+    util::ThreadRename("epgc-wallet");
 
     static bool fOneThread;
     if (fOneThread)
